requirements.txt
Django==5.2.7
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-decouple==3.8
requests==2.32.3
Pillow==11.0.0
pytz==2024.2
django-cors-headers==4.3.1
django-filter==24.2
requests-oauthlib==2.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.115.0
authlib==1.3.0

